Moved admin reset password into the ajax users controller/service. Moved tests
