+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/services/UIUsersService.java
@@ -18,6 +18,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.mail.MailSendException;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
@@ -26,10 +27,12 @@ import ca.corefacility.bioinformatics.irida.exceptions.EntityExistsException;
 import ca.corefacility.bioinformatics.irida.exceptions.EntityNotFoundException;
 import ca.corefacility.bioinformatics.irida.exceptions.InvalidPropertyException;
 import ca.corefacility.bioinformatics.irida.exceptions.PasswordReusedException;
+import ca.corefacility.bioinformatics.irida.model.user.PasswordReset;
 import ca.corefacility.bioinformatics.irida.model.user.Role;
 import ca.corefacility.bioinformatics.irida.model.user.User;
 import ca.corefacility.bioinformatics.irida.repositories.specification.UserSpecification;
 import ca.corefacility.bioinformatics.irida.ria.config.UserSecurityInterceptor;
+import ca.corefacility.bioinformatics.irida.ria.web.exceptions.UIEmailSendException;
 import ca.corefacility.bioinformatics.irida.ria.web.login.PasswordResetController;
 import ca.corefacility.bioinformatics.irida.ria.web.models.tables.TableResponse;
 import ca.corefacility.bioinformatics.irida.ria.web.users.dto.AdminUsersTableRequest;
@@ -39,6 +42,7 @@ import ca.corefacility.bioinformatics.irida.ria.web.users.dto.UserEditRequest;
 import ca.corefacility.bioinformatics.irida.ria.web.utilities.RoleUtilities;
 import ca.corefacility.bioinformatics.irida.service.EmailController;
 import ca.corefacility.bioinformatics.irida.service.ProjectService;
+import ca.corefacility.bioinformatics.irida.service.user.PasswordResetService;
 import ca.corefacility.bioinformatics.irida.service.user.UserService;
 
 import com.google.common.base.Strings;
@@ -55,14 +59,17 @@ public class UIUsersService {
 	private final MessageSource messageSource;
 	private final PasswordEncoder passwordEncoder;
 
+	private final PasswordResetService passwordResetService;
+
 	@Autowired
 	public UIUsersService(UserService userService, ProjectService projectService, EmailController emailController,
-			MessageSource messageSource, PasswordEncoder passwordEncoder) {
+			MessageSource messageSource, PasswordEncoder passwordEncoder, PasswordResetService passwordResetService) {
 		this.userService = userService;
 		this.projectService = projectService;
 		this.emailController = emailController;
 		this.messageSource = messageSource;
 		this.passwordEncoder = passwordEncoder;
+		this.passwordResetService = passwordResetService;
 	}
 
 	/**
@@ -247,6 +254,48 @@ public class UIUsersService {
 	}
 
 	/**
+	 * Create a new {@link PasswordReset} for the given {@link User}
+	 *
+	 * @param userId    The ID of the {@link User}
+	 * @param principal a reference to the logged in user.
+	 * @param locale    a reference to the locale specified by the browser.
+	 * @return text to display to the user about the result of creating a password reset.
+	 * @throws UIEmailSendException if there is an error emailing the password reset.
+	 */
+	public String adminNewPasswordReset(Long userId, Principal principal, Locale locale) throws UIEmailSendException {
+		User user = userService.read(userId);
+		User principalUser = userService.getUserByUsername(principal.getName());
+
+		if (canCreatePasswordReset(principalUser, user)) {
+			try {
+				createNewPasswordReset(user);
+			} catch (final MailSendException e) {
+				throw new UIEmailSendException(
+						messageSource.getMessage("server.password.reset.error.message", null, locale));
+			}
+		} else {
+			throw new UIEmailSendException(
+					messageSource.getMessage("server.password.reset.error.message", null, locale));
+		}
+
+		return messageSource.getMessage("server.password.reset.success.message", new Object[] { user.getFirstName() },
+				locale);
+	}
+
+	/**
+	 * Create a new password reset for a given {@link User} and send a reset password link via email
+	 *
+	 * @param user The user to create the reset for
+	 */
+	private void createNewPasswordReset(User user) {
+		PasswordReset passwordReset = new PasswordReset(user);
+		passwordResetService.create(passwordReset);
+
+		// send a reset password link to user via email
+		emailController.sendPasswordResetLinkEmail(user, passwordReset);
+	}
+
+	/**
 	 * Handle exceptions for the create and update pages
 	 *
 	 * @param ex     an exception to handle
@@ -317,6 +366,32 @@ public class UIUsersService {
 		boolean usersEqual = user.equals(principalUser);
 
 		return usersEqual;
+	}
+
+	/**
+	 * Test if a user should be able to click the password reset button
+	 *
+	 * @param principalUser The currently logged in principal
+	 * @param user          The user being edited
+	 * @return true if the principal can create a password reset for the user
+	 */
+	private boolean canCreatePasswordReset(User principalUser, User user) {
+		Role userRole = user.getSystemRole();
+		Role principalRole = principalUser.getSystemRole();
+
+		if (principalUser.equals(user)) {
+			return false;
+		} else if (principalRole.equals(Role.ROLE_ADMIN)) {
+			return true;
+		} else if (principalRole.equals(Role.ROLE_MANAGER)) {
+			if (userRole.equals(Role.ROLE_ADMIN)) {
+				return false;
+			} else {
+				return true;
+			}
+		}
+
+		return false;
 	}
 
 }
