+++ src/test/java/ca/corefacility/bioinformatics/irida/ria/unit/web/services/UIPasswordResetServiceTest.java
@@ -1,25 +1,19 @@
 package ca.corefacility.bioinformatics.irida.ria.unit.web.services;
 
-import java.security.Principal;
 import java.util.Locale;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.context.MessageSource;
-import org.springframework.ui.ExtendedModelMap;
 
-import ca.corefacility.bioinformatics.irida.model.user.PasswordReset;
-import ca.corefacility.bioinformatics.irida.model.user.Role;
 import ca.corefacility.bioinformatics.irida.model.user.User;
 import ca.corefacility.bioinformatics.irida.ria.web.login.PasswordResetController;
-import ca.corefacility.bioinformatics.irida.ria.web.exceptions.UIEmailSendException;
 import ca.corefacility.bioinformatics.irida.ria.web.services.UIPasswordResetService;
 import ca.corefacility.bioinformatics.irida.service.EmailController;
 import ca.corefacility.bioinformatics.irida.service.user.PasswordResetService;
 import ca.corefacility.bioinformatics.irida.service.user.UserService;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
@@ -39,26 +33,6 @@ public class UIPasswordResetServiceTest {
 		messageSource = mock(MessageSource.class);
 		passwordResetController = new PasswordResetController(passwordResetService);
 		service = new UIPasswordResetService(userService, passwordResetService, emailController, messageSource);
-	}
-
-	@Test
-	void adminNewPasswordResetTest() {
-		User user1 = new User(1L, "Elsa", "elsa@arendelle.ca", "Password1!", "Elsa", "Oldenburg", "1234");
-		user1.setSystemRole(Role.ROLE_ADMIN);
-		User user2 = new User(2L, "Anna", "anna@arendelle.ca", "Password2!", "Anna", "Oldenburg", "5678");
-		Principal principal = () -> user1.getFirstName();
-		String response = null;
-
-		when(userService.read(anyLong())).thenReturn(user2);
-		when(userService.getUserByUsername(anyString())).thenReturn(user1);
-		when(messageSource.getMessage(anyString(), any(), any(Locale.class))).thenReturn("success");
-
-		try {
-			response = service.adminNewPasswordReset(user2.getId(), principal, Locale.ENGLISH);
-		} catch (UIEmailSendException e) {
-			e.printStackTrace();
-		}
-		assertEquals(response, "success");
 	}
 
 	@Test
