+++ src/test/java/ca/corefacility/bioinformatics/irida/ria/unit/web/services/UIUsersServiceTest.java
@@ -14,13 +14,16 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
+import ca.corefacility.bioinformatics.irida.model.user.Role;
 import ca.corefacility.bioinformatics.irida.model.user.User;
+import ca.corefacility.bioinformatics.irida.ria.web.exceptions.UIEmailSendException;
 import ca.corefacility.bioinformatics.irida.ria.web.services.UIUsersService;
 import ca.corefacility.bioinformatics.irida.ria.web.users.dto.UserDetailsModel;
 import ca.corefacility.bioinformatics.irida.ria.web.users.dto.UserDetailsResponse;
 import ca.corefacility.bioinformatics.irida.ria.web.users.dto.UserEditRequest;
 import ca.corefacility.bioinformatics.irida.service.EmailController;
 import ca.corefacility.bioinformatics.irida.service.ProjectService;
+import ca.corefacility.bioinformatics.irida.service.user.PasswordResetService;
 import ca.corefacility.bioinformatics.irida.service.user.UserService;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -35,6 +38,8 @@ public class UIUsersServiceTest {
 	private PasswordEncoder passwordEncoder;
 	private UIUsersService service;
 
+	private PasswordResetService passwordResetService;
+
 	private final User USER1 = new User(1L, "Elsa", "elsa@arendelle.ca", "Password1!", "Elsa", "Oldenburg", "1234");
 	private final User USER2 = new User(2L, "Anna", "anna@arendelle.ca", "Password2!", "Anna", "Oldenburg", "5678");
 
@@ -46,7 +51,8 @@ public class UIUsersServiceTest {
 		emailController = mock(EmailController.class);
 		messageSource = mock(MessageSource.class);
 		passwordEncoder = new BCryptPasswordEncoder();
-		service = new UIUsersService(userService, projectService, emailController, messageSource, passwordEncoder);
+		passwordResetService = mock(PasswordResetService.class);
+		service = new UIUsersService(userService, projectService, emailController, messageSource, passwordEncoder, passwordResetService);
 
 		when(userService.read(anyLong())).thenReturn(USER2);
 		when(userService.getUserByUsername(anyString())).thenReturn(USER1);
@@ -100,5 +106,25 @@ public class UIUsersServiceTest {
 		UserDetailsResponse response = service.changeUserPassword(USER1.getId(), USER1.getPassword(), "Password3!",
 				principal, request);
 		assertEquals(response, expectedResponse, "Received the correct user details response");
+	}
+
+	@Test
+	void adminNewPasswordResetTest() {
+		User user1 = new User(1L, "Elsa", "elsa@arendelle.ca", "Password1!", "Elsa", "Oldenburg", "1234");
+		user1.setSystemRole(Role.ROLE_ADMIN);
+		User user2 = new User(2L, "Anna", "anna@arendelle.ca", "Password2!", "Anna", "Oldenburg", "5678");
+		Principal principal = () -> user1.getFirstName();
+		String response = null;
+
+		when(userService.read(anyLong())).thenReturn(user2);
+		when(userService.getUserByUsername(anyString())).thenReturn(user1);
+		when(messageSource.getMessage(anyString(), any(), any(Locale.class))).thenReturn("success");
+
+		try {
+			response = service.adminNewPasswordReset(user2.getId(), principal, Locale.ENGLISH);
+		} catch (UIEmailSendException e) {
+			e.printStackTrace();
+		}
+		assertEquals(response, "success");
 	}
 }
